WIP: Implement new TopBar
